requirements.txt
google-analytics-data>=0.18.0
google-auth>=2.0.0
anthropic>=0.40.0
python-dotenv>=1.0.0
